Working on report files: writing and reading.
Working on report files: writing and reading.
Functions "Set-ReportsFolder", "Export-WuReport" and "Import-WuReport".
Public function "Get-WuReport" - added timestamps if "PendingReboot" status is detected and / or there are updates ready to be installed.
Also some fixes in the "Get-PendingRebootStatus" function:
Excluded the "Server Manager" registry subkey for non-server SKUs.
Added comment in verbose output about the reasons for the reboot.